+++ src/engraving/libmscore/bracket.h
@@ -80,7 +80,7 @@ public:
     SymId braceSymbol() const { return _braceSymbol; }
     void setBraceSymbol(const SymId& sym) { _braceSymbol = sym; }
     int column() const { return _bi->column(); }
-    int span() const { return _bi->bracketSpan(); }
+    size_t span() const { return _bi->bracketSpan(); }
     qreal magx() const { return _magx; }
 
     System* system() const { return (System*)explicitParent(); }
